Add files via upload
Here we have the benchmark models as presented in the flow framework paper relating to benchmark models, as well as the scenarios which define the RL environment and road networks for each case. Some multi agent environments also specified here, although none which yielded fruitful results; never managed to get the multi-agent environments to converge or achieve meaningful results.